requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
opencv-python==4.5.3.56
face-recognition==1.3.0
numpy==1.24.3
Pillow==9.5.0
python-dotenv==0.19.0
twilio==7.8.0
gunicorn==20.1.0
flask-login==0.5.0
werkzeug==2.3.6
reportlab==3.6.12
openpyxl==3.1.2 
pytest==8.3.3
